requirements.txt
# Requirements pour le Simulateur d'Investissement - CGF GESTION
# Installation : pip install -r requirements.txt

# Framework principal
streamlit>=1.28.0

# Manipulation de données
pandas>=2.0.0
numpy>=1.24.0

# Visualisations
plotly>=5.17.0
altair>=5.1.0

# Images et logos
Pillow>=10.0.0

# PDF generation
reportlab>=4.0.0
matplotlib>=3.7.0

